# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SentinelX API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  active: 'active',
  paused: 'paused',
  archived: 'archived',
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  scope?: string | null;
  status: ProjectStatus;
  assetCount?: number;
  findingCount?: number;
  criticalCount?: number;
  highCount?: number;
  createdAt: string;
  updatedAt: string;
}

export type ProjectInputStatus = typeof ProjectInputStatus[keyof typeof ProjectInputStatus];


export const ProjectInputStatus = {
  active: 'active',
  paused: 'paused',
  archived: 'archived',
} as const;

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  scope?: string;
  status?: ProjectInputStatus;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  active: 'active',
  paused: 'paused',
  archived: 'archived',
} as const;

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  scope?: string;
  status?: ProjectUpdateStatus;
}

export type AssetType = typeof AssetType[keyof typeof AssetType];


export const AssetType = {
  domain: 'domain',
  wildcard: 'wildcard',
  ip_range: 'ip_range',
  url: 'url',
  api: 'api',
} as const;

export type AssetStatus = typeof AssetStatus[keyof typeof AssetStatus];


export const AssetStatus = {
  active: 'active',
  inactive: 'inactive',
  unknown: 'unknown',
} as const;

export interface Asset {
  id: number;
  projectId: number;
  value: string;
  type: AssetType;
  status: AssetStatus;
  /** @nullable */
  notes?: string | null;
  technologies?: string[];
  createdAt: string;
}

export type AssetInputType = typeof AssetInputType[keyof typeof AssetInputType];


export const AssetInputType = {
  domain: 'domain',
  wildcard: 'wildcard',
  ip_range: 'ip_range',
  url: 'url',
  api: 'api',
} as const;

export interface AssetInput {
  /** @minLength 1 */
  value: string;
  type: AssetInputType;
  notes?: string;
  technologies?: string[];
}

export type AssetUpdateType = typeof AssetUpdateType[keyof typeof AssetUpdateType];


export const AssetUpdateType = {
  domain: 'domain',
  wildcard: 'wildcard',
  ip_range: 'ip_range',
  url: 'url',
  api: 'api',
} as const;

export type AssetUpdateStatus = typeof AssetUpdateStatus[keyof typeof AssetUpdateStatus];


export const AssetUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
  unknown: 'unknown',
} as const;

export interface AssetUpdate {
  value?: string;
  type?: AssetUpdateType;
  status?: AssetUpdateStatus;
  notes?: string;
  technologies?: string[];
}

export type FindingSeverity = typeof FindingSeverity[keyof typeof FindingSeverity];


export const FindingSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
  info: 'info',
} as const;

export type FindingStatus = typeof FindingStatus[keyof typeof FindingStatus];


export const FindingStatus = {
  open: 'open',
  in_progress: 'in_progress',
  resolved: 'resolved',
  wont_fix: 'wont_fix',
} as const;

export interface Finding {
  id: number;
  projectId: number;
  /** @nullable */
  assetId?: number | null;
  title: string;
  /** @nullable */
  description?: string | null;
  severity: FindingSeverity;
  status: FindingStatus;
  /** @nullable */
  cvss?: number | null;
  /** @nullable */
  cve?: string | null;
  /** @nullable */
  evidence?: string | null;
  /** @nullable */
  remediation?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type FindingInputSeverity = typeof FindingInputSeverity[keyof typeof FindingInputSeverity];


export const FindingInputSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
  info: 'info',
} as const;

export type FindingInputStatus = typeof FindingInputStatus[keyof typeof FindingInputStatus];


export const FindingInputStatus = {
  open: 'open',
  in_progress: 'in_progress',
  resolved: 'resolved',
  wont_fix: 'wont_fix',
} as const;

export interface FindingInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  severity: FindingInputSeverity;
  status?: FindingInputStatus;
  assetId?: number;
  cvss?: number;
  cve?: string;
  evidence?: string;
  remediation?: string;
}

export type FindingUpdateSeverity = typeof FindingUpdateSeverity[keyof typeof FindingUpdateSeverity];


export const FindingUpdateSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
  info: 'info',
} as const;

export type FindingUpdateStatus = typeof FindingUpdateStatus[keyof typeof FindingUpdateStatus];


export const FindingUpdateStatus = {
  open: 'open',
  in_progress: 'in_progress',
  resolved: 'resolved',
  wont_fix: 'wont_fix',
} as const;

export interface FindingUpdate {
  title?: string;
  description?: string;
  severity?: FindingUpdateSeverity;
  status?: FindingUpdateStatus;
  cvss?: number;
  cve?: string;
  evidence?: string;
  remediation?: string;
}

export type ScanType = typeof ScanType[keyof typeof ScanType];


export const ScanType = {
  recon: 'recon',
  enumeration: 'enumeration',
  vulnerability: 'vulnerability',
  full: 'full',
} as const;

export type ScanStatus = typeof ScanStatus[keyof typeof ScanStatus];


export const ScanStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Scan {
  id: number;
  projectId: number;
  name: string;
  type?: ScanType;
  status: ScanStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress?: number;
  findingsCount?: number;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
  /** @nullable */
  logs?: string | null;
}

export type ScanInputType = typeof ScanInputType[keyof typeof ScanInputType];


export const ScanInputType = {
  recon: 'recon',
  enumeration: 'enumeration',
  vulnerability: 'vulnerability',
  full: 'full',
} as const;

export interface ScanInput {
  /** @minLength 1 */
  name: string;
  type: ScanInputType;
}

export interface DashboardStats {
  totalProjects: number;
  activeProjects: number;
  totalAssets: number;
  totalFindings: number;
  openFindings: number;
  criticalFindings: number;
  highFindings: number;
  runningScans: number;
  completedScans: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  finding_created: 'finding_created',
  scan_completed: 'scan_completed',
  asset_added: 'asset_added',
  project_created: 'project_created',
  finding_updated: 'finding_updated',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  description: string;
  /** @nullable */
  severity?: string | null;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectName?: string | null;
  createdAt: string;
}

export interface SeverityBreakdown {
  critical: number;
  high: number;
  medium: number;
  low: number;
  info: number;
}

export type GetDashboardActivityParams = {
limit?: number;
};