# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SentinelX API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssetStatus } from './assetStatus';
import type { AssetType } from './assetType';

export interface Asset {
  id: number;
  projectId: number;
  value: string;
  type: AssetType;
  status: AssetStatus;
  /** @nullable */
  notes?: string | null;
  technologies?: string[];
  createdAt: Date;
}