# lib/api-zod/src/generated/types/findingInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SentinelX API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FindingInputSeverity } from './findingInputSeverity';
import type { FindingInputStatus } from './findingInputStatus';

export interface FindingInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  severity: FindingInputSeverity;
  status?: FindingInputStatus;
  assetId?: number;
  cvss?: number;
  cve?: string;
  evidence?: string;
  remediation?: string;
}